refactor: unexport remaining internal-only identifiers via white-box tests
Finish unexporting the internal/* identifiers that were exported only because
an external test package named them or because the lowercase name collided with
an import:

- internal/api: GitHub -> gitHub, Keyring -> keyringClient (renamed to avoid the
  keyring import). get_test.go and manager_test.go are converted to white-box
  (package api) so they can reference the unexported names.
- internal/keyring: API -> backendAPI. keyring_test.go is converted to white-box
  (package keyring).
- internal/keyring: delete ParseDate, which had no non-test usage (dead code),
  along with its test.

All changes are inside ghtkn/internal/, so the public API is unchanged.
